requirements.txt
Flask==3.0.3
google-generativeai==0.7.0
requests==2.31.0
pymongo==4.3.3
openai==1.34.0
flask-cors==4.0.1
